requirements.txt
astroid==2.15.5
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
db-dtypes==1.1.1
dill==0.3.6
google-api-core==2.11.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
pandas-gbq==0.19.2
platformdirs==3.5.1
proto-plus==1.22.2
protobuf==4.23.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.0
pylint==2.17.4
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==67.7.2
six==1.16.0
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
